turn MessagePages with assets in DiscussionPages
We can't convert them to the new message system without loosing the asset.

+++ db/migrate/20100603010101_convert_message_page_to_discussion_message.rb
@@ -38,8 +38,12 @@ class ConvertMessagePageToDiscussionMessage < ActiveRecord::Migration
     puts "#{pages.count} Message pages left."
     puts "Converting to Messages."
     pages.each do |page|
-      turn_page_into_messages(page)
-      page.destroy
+      if page.assets.any?
+        page.update_attributes! type: "DiscussionPage"
+      else
+        turn_page_into_messages(page)
+        page.destroy
+      end
     end
   ensure
     enable_timestamps
